Enable Python 3 compatibility
The 'feedformatter' package has been Python 3 compatible for some time
now but this has not made its way to PyPi [1]. Rather than waiting and
hoping for this to be fixed, simply copy the Python 3 code into yasfb,
effectively enabling Python 3 support.

This is mostly a direct copy with a couple of fixes for pep8.

[1] https://github.com/lmaurits/feedformatter/issues/9

Signed-off-by: Stephen Finucane <stephen@that.guru>